+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.0.1] - 2021-08-26
+### Changed
+- Added inputShape as a model param
+### Added
+- Pytorch
+        - Add Roll layer
+        - Add shape to View layer
+
 ## [1.0.0] - 2021-05-27
 ### Changed
 - Updated Metadata in setup.py